Refactor screen navigations to be nested based on feature
