+++ app/src/main/java/com/example/financetracker_app/navigation/extensions/NavControllerExtension.kt
@@ -1,0 +1,9 @@
+package com.example.financetracker_app.navigation.extensions
+
+import androidx.navigation.NavHostController
+
+fun NavHostController.navigateSingleTopTo(route: String) =
+    this.navigate(route) {
+        launchSingleTop = true
+        restoreState = true
+    }
