+++ app/src/main/java/com/example/financetracker_app/navigation/home/HomeScreensDestination.kt
@@ -1,0 +1,12 @@
+package com.example.financetracker_app.navigation.home
+
+import com.example.financetracker_app.navigation.LeafScreenDestination
+import com.example.financetracker_app.navigation.RootScreenDestination
+
+object HomeRoot : RootScreenDestination {
+    override val route = "homeRoot"
+}
+
+object Home : LeafScreenDestination {
+    override val route = "home"
+}
